Disabled X Window System in Cairo installation
Until this commit, we were installing `cairo' without disabling the X
Windows System. That is the default option in the configure step of
`cairo'. However, we are not installing the necessaries libraries (X11)
and headers for that. As a consecuence, when testing in a system where
we don't have these libraries/headers (Ubuntu 14.04), it crashes.

With this commit, we disable the X Windows System option in `cairo'
configure step by setting the option `--with-x=no'.