Updated to TeXLive 2019
Since we just download the binary source of TeXLive, we need to keep it up
to date with the server. So it has been incremented to 2019 (TeXLive 2019
was released April 29th).

A note was also added in the Checklist to keep the users informed on how to
update TeXLive if necessary.