Visual alignment of software version numbers under eachother removed
Until now the software names and versions in
`reproduce/software/config/installation/versions.mk' were placed such that
all the versions were placed under each other. This had a visual effect,
but since many of the software have shorter names, it could sometimes cause
visual confusion on which number belongs to which software.

Since we recently removed a similar situation with the backslashes, with
this commit, there is only a single space between the variable name, `='
and the version number.