Disabled X Window System in netpbm installation for GNU/Linux systems
Until this commit, we were installing `netpbm' without disabling the X
Windows System in GNU/Linux. That is the default option in the configure
step of `netpbm'. However, we are not installing the necessary libraries
(X11) and headers for that. As a consecuence, when testing in a system
where we don't have these libraries/headers (Ubuntu 14.04), it crashes.

With this commit, we disable the X Windows System option in `netpbm' by
giving the answer `none' to one of the last answers to that question in
the configure step.