Removed extra Binutils from GCC's prerequisites
Until this commit, Binutils was brough up as a GCC prerequisite two times:
Once as a special (`host_cc') dependency (for when we want to build GCC),
another time as a dependency in every build. The first one is now obsolete:
on MacOS systems where we don't build Binutils, we define its symbolic
links in the same recipe.

With this commit, we remove the first one, so Binutils is always a
dependency of GCC.