Metastore Git hooks properly written after its installation
Until now, in the recipe to build Metastore, we would store the current
directory in a `current_dir' Bash variable before running `gbuild' to build
Metastore. However, in this Makefile, we aren't using `.ONESHELL'. As a
result, each un-quoted new-line character creates a new shell and the
`current_dir' variable that we used afterwards was empty!

This happened because until recently, Metastore was being built in
`high-level.mk' (where `.ONESHELL' is activated, because we are using GNU
Make to call that Makefile). But in `basic.mk', we are using the host's
Make, which may not be GNU Make, so we can't use any GNU-only features.

Also, this error was hard to notice, because we weren't using `&&' before
writing the final target of Metastore!

With this commit, both this issues are addressed: all the lines in the
recipe are now quoted in the end with a backslash, and the steps before
making the target are separated by `&&', not `;' (which doesn't cause a
crash if the command fails).

This fixes bug #56295.