Shared CFITSIO installation included
By default (when you run `make' within CFITSIO), it only builds a static
library. So until now, the template would only install a static version of
this library.

To be more clear about which libraries are used during the linking of
programs that use CFITSIO, it would be helpful to have shared library
versions also. Also, if the system has problems with static libraries (like
macOS systems), then it won't go looking outside the template to link with
them (these paths are hardcoded in the linker, which we don't build on
macOS and just use the system's tools).

With this commit, we are explicity running `make shared' in the CFITSIO
installation to have shared CFITSIO also. I also removed the old `static'
element (which we don't use anymore in any of the programs and have to
remove some time later!).