Adding SCons into the project
With this commit, SCons has been included into the project. As in the
official webpage of this program says:

  SCons is an Open Source software construction tool—that is, a
  next-generation build tool. Think of SCons as an improved,
  cross-platform substitute for the classic Make utility with integrated
  functionality similar to autoconf/automake and compiler caches such as
  ccache. In short, SCons is an easier, more reliable and faster way to
  build software.

The motivation for including this software into the project is because
there are some software that use this system for the installation. In
particular, the necessity of installing it comes from Imfit. This is an
astronomical software that uses this system so that is why it has been
included into the project. In principle it is simple, and I have tested
it on Mac OS sytem without any problem.