Added Imfit into the project
With this commit, `imfit' has been included into the project. As described
in the official webpage (http://www.mpe.mpg.de/~erwin/code/imfit/):

  Imfit is a program for fitting astronomical images -- especially images
  of galaxies, though it can in principle be used for fitting other
  sources.

This program uses the `scons' program to build itself (not the traditional
GNU `configure' and `make' system). I have tested it by using an example
that can be found in the `examples' directory of the decompressed
tarball. It works fine without any error on my Mac OS system.