Corrected typo in Imfit installation, but not working yet
With this commit, I have corrected a typo in the installation of
`imfit'. In principle it is working on Mac OS system but when I tried to
install it in GNU/Linux system it crashed. It complains about a problem
with the `cfitsio' library:
  /pathto/libcfitsio.a: error adding symbols: File format not recognized

As in the last commits done by Mohammad in the project, `cfitsio'
installation has been upgraded and also the installation has been
modified, I am going to try this new version. I will also check if it is
a problem about dynamic/static library.

It also seems that the installation is taking libraries and variables
from the host system, so I will check that everything used is from our
own programs/libraries.