Building single and double precission float libraries of FFTW
Until this commit, `fftw' was building single precission float library
because `scamp' had problem on using the default double precission
library. However,`imfit' and maybe other programs will use double
precission float libraries.

With this commit, `fftw' installation is done twice in order to build
single and double precission float libraries. It does not complain and
goes up to the end without crashing. However, more tests are needed in
order to ensure that there is not problem on having both libraries
because in principle, they share the same header file and that could
cause problems.