TeXlive: allows for updating itself and removes old version file
Until now, TeXlive would only attempt to download/update LaTeX packages,
but after some time, it will be necessary for it to update itself. When
this happens, it complains about running it with `update --self'.

To fix this problem, before installing/updating the LaTeX packages, we will
first update tlmgr in any case.

Also, we weren't removing the TeXlive package version file before appending
the values to it. So it was just repeating the packages every time it was
updated. With this commit, it is being removed on every new build.

Finally, the version of Git was updated to its most recent version.