Metastore now has rpath
Until now we hadn't noticed that Metastore's build doesn't include an
`rpath'. As a result I got a crash during a Git commit.

With this commit, on GNU/Linux systems, we now add rpath to the installed
Metastore executable. During the build, I also cleaned up the Metastore
build recipe to be more clear and readable (in particular we only check the
user name and group name when metastore is actually built).