Binutils is built alone, before GCC
Until now, on GNU/Linux systems, GNU Binutils was built in parallel with
other programs. As a result, the moment that the Binutils executables (most
importantly `ld') are being installed, if another program is using them, we
can have a crash.

For some reason, this crash doesn't happen on Fedora or Arch Linux, but on
some Debian-based distors, it causes the following crash when we are
building Git (which happened to be built while Binuntils is being
installed):

credential-cache--daemon.o: file not recognized: File format not recognized

In a following run of `./configure' (when Binutils has been installed), Git
will build and install successfully.

To fix this problem, with this commit, all basic programs are now a
prerequisite of Binutils and Binutils is the sole prerequisite of GCC
(which is the final target of `basic.mk').

Also, the GCC configure options were re-ordered to be easier to read (the
`--with-*' options under each other, the `--enable-*' options under each
other and so on).