Removing libiconv before building pkg-config
An existing `libiconv' can cause a conflict with `pkg-config', this is why
`libiconv' depends on `pkg-config'. On a clean build, `pkg-config' is built
first. But when we don't have a clean build (and `libiconv' exists) there
will be a problem.

With this commit, before re-building `pkg-config', we'll remove any
possibly existing installation of `libiconv'.