C and Fortran compiler environment set in high-level.mk
Until now, we weren't explicitly setting the C and Fortran compiler
environment variables (`CC' and `F77'). As a result, if the user's system
had already set them, there would be a problem (and the system's compilers
would be used).

With this commit, we are explicitly setting these two environment variables
at the start of `high-level.mk'.

This bug was found after a discussion with Elham Saremi.