libffi always installed in the lib directory
On some Fedora systems, libffi installs under `lib64', not `lib'. As a
result, Python's Setuptools can't find it and will not built (complaining
about not finding the `_ctypes' module).

With this commit, we fix this problem by explicitly putting a copy of
libffi's installed libraries within the `lib' directory.

This issue was found while testing the pipeline with Elham Saremi and Hamed
Altafi.