Compiler checks added to the configure script
To build the software, we need a C/C++ compiler. But until now at configure
time there was no check on this. As a result, when the system didn't have
them, the user would only learn in a crash of the programs.

With this commit, some checks have been added at configure time to do basic
checks and inform the user if necessary.