Minor corrections in high-level programs
While doing a clean build, several issues were found in the pipeline and
corrected. The main issue was that with the recent installation of
`libiconv', the GCC standard C++ library depends on `libiconv', so we need
to explicity add an `-liconv' to any C++ compilation.

The other corrected points are:
 - The C++ compiler is now explicitly defined in `CXX'.
 - libgit2 and WCSLIB's recipes weren't using `&&' between statements, so
   if there was an error, it would still build the target!
 - The CMake bootstrapping script now builds much faster in parallel.