Better explanation of suggested commit messages in REAME-hacking.md
Until now the description of the commit message guidelines wasn't clear
enough and could cause confusion, in particular because it didn't describe
why some basic formatting issues are mandatory.

With this commit, it is explained that the main reason we require
contributors for follow this format is "consistency" within the
project. Also generally it was edited to become more elaborate and explain
the points more clearly.

I also ran a spell check over the whole file and fixed a few typos.

This correction was suggested by Mohammad-reza Khellat.