Corrected typo in environment before running make
We recently moved the system's `rm' program absolute address to a shell
variable that is found during the `./project' script. But I had forgot to
account for the difference between the Make and Bash variable naming
differences. I had also forgot to add a value to the HOME variable.

With this commit both are corrected: the system's `rm' path is now called
`sys_rm' and the HOME variable is set.