Tip added at initial configuration notice on how to see progress
The configuration step (building all the ncessary software) can take some
time. It is natual for the user to want to see how the build is going
(which software is being built at every moment). So far, we have only put a
"Inspecting status" section in `README-hacking.md' that describes a
solution, but some early users may not have read it yet.

With this commit a short tip was added in the initial installation notice
to inform the user of this very useful command.