OpenSSH necessary to use OpenMPI
Until now, OpenMPI was being installed without any dependency. This was
fine because it would indeed build. But the moment you tried loading
something that depends on it (for example `mpi4py' through `astropy'), you
would get an error complaining that SSH isn't present.

With this commit, the pipeline now also installs OpenSSH to solve this
problem.