WCSLIB 6.4 doesn't need explicit link with CFITSIO
This was a bug in WCSLIB 6.3 that has been fixed in WCSLIB 6.4. From
WCSLIB's changelog: "The rule change to the Fortran makefile in v6.3 to add
getwcstab_f.o to the sharable library causes it to depend on CFITSIO to
resolve fits_get_wcstab(). Hence backed out of that change.".

The actual error was like this:

       Undefined symbols for architecture x86_64:
         "_fits_read_wcstab", referenced from:
             _ftwcst_ in getwcstab_f.o
         "_gFitsFiles", referenced from:
             _ftwcst_ in getwcstab_f.o
       ld: symbol(s) not found for architecture x86_64