Using libgit2 implementation of libconv with the option -DUSE_ICONV=OFF
When trying to install `libgit2' on my macOS system, it complains  about
not finding `_iconv*' functions! But apparently `libgit2' has its own
implementation of `libiconv' that it uses if it can't find `libiconv' on
macOS. So, the solution to this problem was to add `-DUSE_ICONV=OFF' to
the configuration options of `libgit2'.

I reported this issue that now is fixed thanks to the help of Mohammad
Akhlaghi.